feat: Model Select View 추가
